rename clients -> interfaces
