+++ tests/messages_query.rs
@@ -5,10 +5,10 @@
 
 use std::io::Cursor;
 
-use dwn_node::clients::grants::GrantBuilder;
-use dwn_node::clients::messages::{QueryBuilder, ReadBuilder};
-use dwn_node::clients::protocols::ConfigureBuilder;
-use dwn_node::clients::records::{Data, ProtocolBuilder, WriteBuilder};
+use dwn_node::interfaces::grants::GrantBuilder;
+use dwn_node::interfaces::messages::{QueryBuilder, ReadBuilder};
+use dwn_node::interfaces::protocols::ConfigureBuilder;
+use dwn_node::interfaces::records::{Data, ProtocolBuilder, WriteBuilder};
 use dwn_node::messages::MessagesFilter;
 use dwn_node::permissions::Scope;
 use dwn_node::protocols::Definition;
