+++ tests/protocols_query.rs
@@ -7,8 +7,8 @@ use std::time::Duration;
 
 use base64ct::{Base64UrlUnpadded, Encoding};
 use chrono::Days;
-use dwn_node::clients::grants::{GrantBuilder, RevocationBuilder};
-use dwn_node::clients::protocols::{ConfigureBuilder, QueryBuilder};
+use dwn_node::interfaces::grants::{GrantBuilder, RevocationBuilder};
+use dwn_node::interfaces::protocols::{ConfigureBuilder, QueryBuilder};
 use dwn_node::permissions::Scope;
 use dwn_node::protocols::{Definition, ProtocolType, ProtocolsFilter};
 use dwn_node::{Error, Method, cid, endpoint};
