+++ tests/records_query.rs
@@ -3,8 +3,8 @@
 use std::io::Cursor;
 
 use chrono::{DateTime, Duration, Utc};
-use dwn_node::clients::protocols::ConfigureBuilder;
-use dwn_node::clients::records::{Data, ProtocolBuilder, QueryBuilder, WriteBuilder};
+use dwn_node::interfaces::protocols::ConfigureBuilder;
+use dwn_node::interfaces::records::{Data, ProtocolBuilder, QueryBuilder, WriteBuilder};
 use dwn_node::protocols::Definition;
 use dwn_node::records::{RecordsFilter, Sort};
 use dwn_node::store::{MAX_ENCODED_SIZE, Pagination};
